requirements.txt
Pillow~=10.0.1
pygame~=2.5.2
tk = 0.1.0
pyglet~=2.0.15